requirements.txt
# PipeZone - Python Dependencies
# These are installed in the Jupyter Docker container

# Core dependencies
pyyaml==6.0.1
python-dotenv==1.0.1

# Database drivers - PostgreSQL
psycopg2-binary==2.9.9

# Database drivers - MySQL/MariaDB
mysql-connector-python==8.2.0
pymysql==1.1.0
mysqlclient==2.2.1

# Database drivers - SQL Server
pyodbc==5.0.1
pymssql==2.2.11

# Database drivers - Oracle
cx-Oracle==8.3.0
oracledb==2.0.1

# Database drivers - MongoDB
pymongo==4.6.1
motor==3.3.2

# Database drivers - Cassandra
cassandra-driver==3.29.0

# Database drivers - Redis
redis==5.0.1
hiredis==2.3.2

# Database drivers - Elasticsearch
elasticsearch==8.11.1

# Database drivers - ClickHouse
clickhouse-driver==0.2.7
clickhouse-connect==0.7.0

# Database drivers - Snowflake
snowflake-connector-python==3.6.0
snowflake-sqlalchemy==1.5.1

# Object Storage
boto3==1.34.24
minio==7.2.3
s3fs==2024.2.0

# Secrets Management
hvac==2.1.0

# SQLAlchemy
sqlalchemy==2.0.25
alembic==1.13.1
sqlparse==0.4.4

# Data processing
pandas==2.1.4
numpy==1.26.3
pyarrow==14.0.2
fastparquet==2024.2.0

# Data validation
great-expectations==0.18.8
pandera==0.18.0

# Airflow client
apache-airflow-client==2.8.0

# Utilities
requests==2.31.0
tenacity==8.2.3
tqdm==4.66.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
